split s0 and dallas code
