+++ HeishaMon/version.h
@@ -1,1 +1,1 @@
-static const char* heishamon_version = "0.9b-iy-5-LittleFS";
+static const char* heishamon_version = "0.9b-iy-6-LittleFS";
